Refactor address-related controllers and views
This commit refactors the address-related controllers and views by creating a new `AddAddressController` abstract class and renaming the existing `AddAddressPart1Controller` to `AddAddressController`. It also adds a new `EditAddressController` and `EditAddressDetailsController` to handle editing of existing addresses. The `ViewAddressController` has been updated to include a new `goToEditAddress` method to navigate to the `EditAddress` screen. Additionally, the `AdressData` class has been updated to convert latitude and longitude values to strings before sending them to the server. Finally, the `EditAddress` and `EditAddressDetails` screens have been implemented with their respective controllers.