Update Firebase dependencies and configuration
This commit updates the Firebase dependencies for `firebase_core` and `firebase_messaging` to versions `2.32.0` and `14.9.4`, respectively. It also updates the `google-services` plugin version to `4.3.15` and adds the necessary Firebase configuration files. Additionally, it subscribes users to the `users` topic upon successful login and adds a PHP function for sending GCM messages.