feat: Add refresh functionality to pending orders page and update Jiffy package version
This commit adds a refresh function to the pending orders page in the `PendingConrollerImp` class. It also updates the Jiffy package version to `6.3.1` in the `pubspec.yaml` file. Additionally, the `NotificationServices` class now calls the `refreshOrderPindingPage` function when a notification with the `pagename` of `orderpendingrefresh` is received, which triggers the refresh function on the pending orders page. The `CustomCard` widget now displays the order date and time using the Jiffy package's `fromNow` method.