Add goToAddAddress method to CheckoutControllerImp
This commit adds a new method called goToAddAddress to the CheckoutControllerImp class. This method navigates to the add address screen when called. Additionally, the commit also updates the chooseDeliveryType method to set the addressId to the first address in the viewAddressList if the delivery type is 0 and the list is not empty. Finally, the commit updates the statusRequest value to none if the response from the server is not successful.